requirements.txt
numpy
pandas
duckduckgo-search
yfinance
google-generativeai
phi
python-dotenv